Move SignalTypes to util
